requirements.txt
Django==4.0.2
djangorestframework==3.13.1
django-environ==0.8.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
django-translated-fields==0.11.3
Pillow==9.0.1
requests==2.27.1